feat:系统管理-角色管理-权限设置、菜单编辑回显

+++ common/common-util/src/main/java/com/autumn/blog/common/base/TreeAble.java
@@ -13,9 +13,9 @@ public interface TreeAble<T> {
 
     Object getPid();
 
-    Long getDeep();
+    Integer getDeep();
 
-    Long getSort();
+    Integer getSort();
 
     List<T> getChildren();
